Refactor sprokit pipeline construction.
This part of the process_util consisted of many classes all in the
same file. This change splits all these classes into separate
files, header and implementation.
