+++ sprokit/src/sprokit/pipeline_util/check_provider.h
@@ -1,0 +1,64 @@
+/*ckwg +29
+ * Copyright 2016 by Kitware, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither name of Kitware, Inc. nor the names of any contributors may be used
+ *    to endorse or promote products derived from this software without specific
+ *    prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE AUTHORS OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+/**
+ * @file   check_provider.h
+ * @brief  Interface to check_provider class
+ */
+
+#ifndef SPROKIT_PIPELINE_UTIL_CHECK_PROVIDER_H
+#define SPROKIT_PIPELINE_UTIL_CHECK_PROVIDER_H
+
+#include "bakery_base.h"
+
+#include <vital/config/config_block.h>
+
+#include <boost/variant.hpp>
+
+
+namespace sprokit {
+
+class check_provider
+  : public boost::static_visitor<bool>
+{
+  public:
+    check_provider(config_provider_t const& provider);
+    ~check_provider();
+
+    bool operator () (kwiver::vital::config_block_value_t const& value) const;
+    bool operator () (bakery_base::provider_request_t const& request) const;
+  private:
+    config_provider_t const m_provider;
+};
+
+
+} // end namespace sprokit
+
+#endif /* SPROKIT_PIPELINE_UTIL_CHECK_PROVIDER_H */
