+++ sprokit/src/sprokit/pipeline_util/ensure_provided.cxx
@@ -1,0 +1,75 @@
+/*ckwg +29
+ * Copyright 2016 by Kitware, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither name of Kitware, Inc. nor the names of any contributors may be used
+ *    to endorse or promote products derived from this software without specific
+ *    prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE AUTHORS OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+/**
+ * @file   ensure_provided.cxx
+ * @brief  Implementation for ensure_provided class.
+ */
+
+#include "ensure_provided.h"
+#include "pipe_bakery_exception.h"
+
+
+namespace sprokit {
+
+// ------------------------------------------------------------------
+ensure_provided
+::ensure_provided()
+{
+}
+
+
+ensure_provided
+::~ensure_provided()
+{
+}
+
+
+// ------------------------------------------------------------------
+kwiver::vital::config_block_value_t
+ensure_provided
+::operator () (kwiver::vital::config_block_value_t const& value) const
+{
+  return value;
+}
+
+
+// ------------------------------------------------------------------
+kwiver::vital::config_block_value_t
+ensure_provided
+::operator () (bakery_base::provider_request_t const& request) const
+{
+  config_provider_t const& provider = request.first;
+  kwiver::vital::config_block_value_t const& value = request.second;
+
+  throw unrecognized_provider_exception("(unknown)", provider, value);
+}
+
+} // end namespace sprokit
