+++ sprokit/src/sprokit/pipeline_util/loaded_cluster.h
@@ -1,0 +1,65 @@
+/*ckwg +29
+ * Copyright 2016 by Kitware, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither name of Kitware, Inc. nor the names of any contributors may be used
+ *    to endorse or promote products derived from this software without specific
+ *    prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE AUTHORS OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+/**
+ * @file   loaded_cluster.h
+ * @brief  Interface to loaded_cluster class.
+ */
+
+#ifndef SPROKIT_PIPELINE_UTIL_LOADED_CLUSTER_H
+#define SPROKIT_PIPELINE_UTIL_LOADED_CLUSTER_H
+
+#include <sprokit/pipeline/process_cluster.h>
+
+
+namespace sprokit {
+
+// ----------------------------------------------------------------
+/**
+ * @brief Friendly process cluster.
+ *
+ * This class provides friend access to process_cluster by
+ * cluster_creator.
+ *
+ */
+class loaded_cluster
+  : public process_cluster
+{
+  public:
+    loaded_cluster(kwiver::vital::config_block_sptr const& config);
+    ~loaded_cluster();
+
+    friend class cluster_creator;
+};
+
+} // end namespace sprokit
+
+
+#endif /* SPROKIT_PIPELINE_UTIL_LOADED_CLUSTER_H */
